project/requirements.txt
asgiref==3.5.2
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.4
defusedxml==0.7.1
Django==4.1.3
django-allauth==0.51.0
django-cors-headers==3.13.0
django-filter==22.1
django-widget-tweaks==1.4.12
djangorestframework==3.14.0
idna==3.4
oauthlib==3.2.2
pycparser==2.21
PyJWT==2.6.0
python3-openid==3.2.0
pytz==2022.6
requests==2.31.0
requests-oauthlib==1.3.1
sqlparse==0.4.3
tzdata==2022.6
urllib3==1.26.13
line-bot-sdk==3.5.0
openai==0.27.0
SpeechRecognition==3.10.3
pydub==0.25.1

